# openapi.generated2.yaml
openapi: 3.1.0
info:
  title: FastAPI
  version: 0.1.0
paths:
...
  /api/v1/transforms/suggest:
    post:
      tags:
      - Transforms
      summary: Suggest transforms
      description: accepts schemas/placeholder_probe.schema.json; returns schemas/suggest_response.schema.json
      operationId: post_transforms_suggest_api_v1_transforms_suggest_post
      parameters:
      - name: If-Match
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: If-Match
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '428':
          content:
            application/problem+json: {}
          description: Precondition Required
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/transforms/preview:
    post:
      tags:
      - Transforms
      summary: Preview transforms
      description: returns schemas/transforms_preview_response.schema.json
      operationId: post_transforms_preview_api_v1_transforms_preview_post
      parameters:
      - name: If-Match
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: If-Match
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '428':
          content:
            application/problem+json: {}
          description: Precondition Required
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError